WIP: refactor Ellipsis to usePageItemControl
