+++ apps/app/src/components/Sidebar/PageTreeItem/Ellipsis.module.scss
@@ -4,6 +4,6 @@
     // left: 0;
     // display: inline-flex;
 
-    // width: calc(100% - 80px);
+    width: calc(100% - 35px);
   }
 }
