+++ apps/app/src/components/Sidebar/PageTreeItem/index.ts
@@ -1,2 +1,1 @@
 export * from './PageTreeItem';
-export * from './Ellipsis';
